Various fixes:
handle electrum server disconnects and switch to new servers automatically, refactor server lists to data file

add SSL connections to electrum

Bugfixes in electrum code:
Restart on connectionFailed as well as connectionLost
Query utxo set must return (only) None for nonexistent/spent utxos.
Dont include twisted logging for wallet-tool

start thread immediately for early blockchain queries

timeout non-responsive servers

fix abstractBCI refs in subclasses
