+++ jmclient/jmclient/blockchaininterface.py
@@ -202,6 +202,13 @@ class ElectrumWalletInterface(BlockchainInterface): #pragma: no cover
     def add_tx_notify(self, txd, unconfirmfun, confirmfun, notifyaddr):
         log.debug("Dummy electrum interface, no add tx notify")
 
+    def outputs_watcher(self, wallet_name, notifyaddr,
+                        tx_output_set, uf, cf, tf):
+        log.debug("Dummy electrum interface, no outputs watcher")
+
+    def tx_watcher(self, txd, ucf, cf, sf, c, n):
+        log.debug("Dummy electrum interface, no tx watcher")
+
     def pushtx(self, txhex, timeout=10):
         #synchronous send
         from electrum.transaction import Transaction
